working DAG construction and alignment
The current implementation is passing a number of test cases in
simulated data.

    - Fixed p5 node ordering bug (caused construction-dependent
      alignments across multiple nodes).
    - Added documentation of command-line options.
    - Improved output format.

Default output is now a primitive graph-CIGAR describing the alignment
across the nodes of the DAG.  The cigar format is:

    [node_name]:[x,y];[cigar]| ...

Currently node names reflect the origin of the sequence data.  ref.* are
reference nodes, and alt.* are alternates from the VCF.  Reference nodes
from the fasta input are of the form ref.r, while those from the VCF are
in the form ref.0, and alternates are of the form alt.N, where N is the
allele order in the VCF (1=1st, 2=2nd, etc.).  The cigar is in expanded
form with one char per base in read.  "|" delimits entries.