major cleanup, now usable
Lots of debugging flags were pulled out.

The check for alignment acceptibility has been refined to allow acceptance only
when the sum of mismatches and softclip qualities is >= before realignmnet than
after, or if gaps are introduced the strictly reduce this sum, implying better
alignment to the graph.  If gaps are reduced while overall qsum of mismatches
and softclips does not increase, the realignment is accepted.